Refactor: util format 폴더 구조 정리
