+++ src/components/coupon-registration/discount-type/index.tsx
@@ -10,7 +10,10 @@ import {
   RATE_DISCOUNT_TYPE,
 } from '@/constants/coupon-registration';
 import { useEffect, useState } from 'react';
-import { numberFormat, removeNumberFormat } from '@/utils/Format/numberFormat';
+import {
+  numberFormat,
+  removeNumberFormat,
+} from '@/utils/format-fn/numberFormat';
 import { InputChangeEvent, MouseEvent } from '@/types/event';
 import { isNumber } from '@/utils/is-number';
 import { handleEnterKeyDown } from '@/utils/keydown/handleEnterKeyDown';
