+++ src/components/point-detail/point-box/index.tsx
@@ -10,7 +10,7 @@ import {
 } from '@stores/point-detail/atoms';
 import { useRecoilState } from 'recoil';
 
-import { numberFormat } from '@/utils/Format/numberFormat';
+import { numberFormat } from '@/utils/format-fn/numberFormat';
 const MONTH_JANUARY = 1;
 const MONTH_DECEMBER = 12;
