+++ src/utils/format-fn/dateFormat.ts
@@ -30,3 +30,18 @@ export const getChartDate = (index: number) => {
   )}/${String(pastDate.getDate()).padStart(targetLength, fillChar)}`;
   return formattedDate;
 };
+
+export const orderDate = () => {
+  const currentDate = new Date();
+
+  const year = currentDate.getFullYear();
+  const month = (currentDate.getMonth() + 1).toString().padStart(2, '0');
+  const day = currentDate.getDate().toString().padStart(2, '0');
+  const hours = currentDate.getHours().toString().padStart(2, '0');
+  const minutes = currentDate.getMinutes().toString().padStart(2, '0');
+  const seconds = currentDate.getSeconds().toString().padStart(2, '0');
+
+  const orderNumber = `O-${year}${month}${day}${hours}${minutes}${seconds}`;
+
+  return orderNumber;
+};
